FIX: many warnings
(cherry picked from commit 0a59ad3759cde23c6aa4dc7c114473e95d5cd1b1)

 Conflicts:
	SofaKernel/modules/SofaBaseVisual/SofaBaseVisual_test/VisualModelImpl_test.cpp
	SofaKernel/modules/SofaSimulationCommon/SceneLoaderPHP.cpp
	SofaKernel/modules/SofaSimulationCommon/SceneLoaderXML.cpp
	applications/plugins/SofaHAPI/initSofaHAPI.cpp
	applications/plugins/SofaPython/initSofaPython.cpp
	modules/SofaMiscForceField/WashingMachineForceField.cpp

+++ modules/SofaGeneralEngine/SofaGeneralEngine_test/TransformEngine_test.cpp
@@ -67,7 +67,7 @@ namespace
 
     template<class DataTypes>
     void setRot(typename DataTypes::Coord& coord, const sofa::helper::Quater<SReal>& rot)
-    { setRotWrapper<DataTypes, DataTypes::Coord::spatial_dimensions, DataTypes::Coord::total_size == DataTypes::Coord::spatial_dimensions>::setRot(coord, rot); }
+    { setRotWrapper<DataTypes, DataTypes::Coord::spatial_dimensions, (int)DataTypes::Coord::total_size == (int)DataTypes::Coord::spatial_dimensions>::setRot(coord, rot); }
 
     /// Create a coord of the specified type from a Vector3 and a Quater
     template<class DataTypes>
